@@ -101,7 +101,7 @@ import { AppendToInputDirective } from './directives/append-to-input.directive';
 })
 export class NgxMaterialTimepickerModule {
 
-    static setLocale(locale: string): ModuleWithProviders {
+    static setLocale(locale: string): ModuleWithProviders<NgxMaterialTimepickerModule> {
         return {
             ngModule: NgxMaterialTimepickerModule,
             providers: [
