@@ -1,9 +1,8 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { NO_ERRORS_SCHEMA } from '@angular/core';
 import { NgxMaterialTimepicker24HoursFaceComponent } from './ngx-material-timepicker-24-hours-face.component';
-import * as TimepickerTime from '../../utils/timepicker-time.utils';
 import { DateTime } from 'luxon';
-import { spyOnFunction } from '../../ngx-material-timepicker.component.spec';
+import { TimepickerTimeUtils } from '../../utils/timepicker-time.utils';
 
 describe('NgxMaterialTimepicker24HoursFaceComponent', () => {
     let fixture: ComponentFixture<NgxMaterialTimepicker24HoursFaceComponent>;
@@ -19,10 +18,10 @@ describe('NgxMaterialTimepicker24HoursFaceComponent', () => {
     });
 
     it('should call disableHours', () => {
-        const spy = spyOnFunction(TimepickerTime, 'disableHours');
+        const spy = spyOn(TimepickerTimeUtils, 'disableHours');
         const time = DateTime.fromJSDate(new Date());
         const format = 24;
-        const hours = TimepickerTime.getHours(format);
+        const hours = TimepickerTimeUtils.getHours(format);
 
         component.minTime = time;
         component.maxTime = time;
